Added test_sims dir for testing and a python single module test.
The test_sims dir in cpp/ is meant to hold simulation definitons that
pair with specific simulation test cases.  These will be further
organized into subdirectories, the first of which is the
single_module_tests directory which right now contains a simulation
definition for testing whether the theoretical period of a single module
matches the period in the simulation.

The tests dir in python is for scripts that use the test simulation
results in cpp/test_sims.  The first of these is
single_module_theory_period.py which looks at a single module simulation
and compares the analytical period to the simulated period.  Energy is
also plotted.